# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: API responsible for mananging Users and their Games
  version: 1.0.0
  title: Queens Gambit API
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: localhost:8080
basePath: /queens-gambit-api/v1
tags:
  - name: User
    description: Operations regarding the users
  - name: Game
    description: Operations regarding the games
schemes:
  - http
paths:
  /games/{fullName}:
    get:
      tags:
        - Game
      summary: "Gets data from Game by its name"
      description: "Returns a single game"
      operationId: "getGamesByFullName"
      produces:
        - "application/json"
      parameters:
        - name: "fullName"
          in: "path"
          description: "FullName of Game to return"
          required: true
          type: string
      responses:
        '200':
          description: "Game data suscessfully retrieved"
          schema:
            $ref: "#/definitions/GameResponse"
        '500':
          description: "Internal error ocurred while getting Game data"
        '400':
          description: "Invalid Request"
        '404':
          description: "No Games found"

  /users/login:
    post:
      tags:
        - User
      summary: Log the User in
      operationId: loginUser
      consumes:
        - "application/json"
      parameters:
        - in: body
          name: user
          description: "User that is being logged-in"
          required: true
          schema:
            $ref: '#/definitions/UserLoginRequest'
      responses:
        '200':
          description: User suscessfully logged in
        '500':
          description: Internal error ocurred while login user
        '400':
          description: Invalid User Login Request Provided

  /users/{login}:
    post:
      tags:
        - User
      summary: Creates an User
      operationId: saveUser
      consumes:
        - "application/json"
      parameters:
        - in: body
          name: user
          description: "User that is being saved"
          required: true
          schema:
            $ref: '#/definitions/UserRequest'
      responses:
        '200':
          description: User data suscessfully saved
        '500':
          description: Internal error ocurred while saving user
        '400':
          description: Invalid User Request Provided

    get:
      tags:
        - User
      summary: "Gets data from User by login"
      description: "Returns a single user"
      operationId: "getUserByLogin"
      produces:
        - "application/json"
      parameters:
        - name: "login"
          in: "path"
          description: "Login of User to return"
          required: true
          type: string
      responses:
        '200':
          description: "User data suscessfully retrieved"
          schema:
            $ref: "#/definitions/UserResponse"
        '500':
          description: "Internal error ocurred while getting User data"
        '400':
          description: "Invalid Request"
        '404':
          description: "No Users found"

definitions:
  GameResponse:
    type: object
    properties:
      full_name:
        type: string
        description: Full name of the game
      launch_year:
        type: integer
        format: int64
        description: Launch year
      developer:
        type: string
        format: int64
        description: Developer of the game
      players:
        type: array
        description: List of players
        items:
          $ref: '#/definitions/UserReducedResponse'
      owners:
        type: array
        description: List of owners
        items:
          $ref: '#/definitions/UserReducedResponse'

  UserRequest:
    type: object
    required:
      - login
      - full_name
      - city_name
      - age
      - cell_phone
      - available_in
      - password
    properties:
      login:
        type: string
        description: User login
        minLength: 1
      full_name:
        type: string
        format: int64
        description: Full name of the User
      city_name:
        type: string
        description: Ip Address of the Machine
        minLength: 1
      age:
        type: integer
        format: int64
        description: User age
      cell_phone:
        type: string
        description: User cellphone
        minLength: 1
      home_phone:
        type: string
        description: User homephone
      insta:
        type: string
        description: User instagram
      face:
        type: string
        description: User facebook
      twitter:
        type: string
        description: User twitter
      available_in:
        type: string
        description: User availability
        minLength: 1
      password:
        type: string
        description: User password
        minLength: 1

  UserLoginRequest:
    type: object
    required:
      - login
      - password
    properties:
      login:
        type: string
        description: User login
        minLength: 1
      password:
        type: string
        description: User password
        minLength: 1

  UserResponse:
    type: object
    properties:
      login:
        type: string
        description: User login
        minLength: 1
      full_name:
        type: string
        format: int64
        description: Full name of the User
      city_name:
        type: string
        description: Ip Address of the Machine
      age:
        type: integer
        format: int64
        description: User age
      cell_phone:
        type: string
        description: User cellphone
      home_phone:
        type: string
        description: User homephone
      insta:
        type: string
        description: User instagram
      face:
        type: string
        description: User facebook
      twitter:
        type: string
        description: User twitter
      created_at:
        type: string
        description: User date_of_creation
      available_in:
        type: string
        description: User availability
      user_wanted_games:
        type: array
        description: List of wanted games
        items:
          $ref: '#/definitions/ReducedGameResponse'
      user_owned_games:
        type: array
        description: List of owned games
        items:
          $ref: '#/definitions/ReducedGameResponse'
      user_friends:
        type: array
        description: List of friendships
        items:
          $ref: '#/definitions/UserReducedResponse'

  UserReducedResponse:
    type: object
    properties:
      full_name:
        type: string
        description: name of the player
      login:
        type: string
        description: login of the player

  ReducedGameResponse:
    type: object
    properties:
      full_name:
        type: string
        description: Full name of the game